rename foo -> 3d
