+++ app/upgrades/drs-2/upgrade.go
@@ -4,6 +4,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/module"
 	upgradetypes "github.com/cosmos/cosmos-sdk/x/upgrade/types"
+	"github.com/dymensionxyz/rollapp-evm/app/upgrades"
 	evmkeeper "github.com/evmos/evmos/v12/x/evm/keeper"
 
 	rollappparamskeeper "github.com/dymensionxyz/dymension-rdk/x/rollappparams/keeper"
@@ -11,11 +12,12 @@ import (
 )
 
 func CreateUpgradeHandler(
+	kk upgrades.UpgradeKeepers,
 	mm *module.Manager,
 	configurator module.Configurator,
 ) upgradetypes.UpgradeHandler {
 	return func(ctx sdk.Context, _ upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
-		if err := HandleUpgrade(ctx, rpKeeper, evmKeeper); err != nil {
+		if err := HandleUpgrade(ctx, kk.RpKeeper, kk.EvmKeeper); err != nil {
 			return nil, err
 		}
 		return mm.RunMigrations(ctx, configurator, fromVM)
