+++ app/upgrades/drs-5/upgrade.go
@@ -5,18 +5,19 @@ import (
 	"github.com/cosmos/cosmos-sdk/types/module"
 	upgradetypes "github.com/cosmos/cosmos-sdk/x/upgrade/types"
 	rollappparamskeeper "github.com/dymensionxyz/dymension-rdk/x/rollappparams/keeper"
+	"github.com/dymensionxyz/rollapp-evm/app/upgrades"
 	evmkeeper "github.com/evmos/evmos/v12/x/evm/keeper"
 
 	drs4 "github.com/dymensionxyz/rollapp-evm/app/upgrades/drs-4"
 )
 
 func CreateUpgradeHandler(
-	kk UpgradeKeepers,
+	kk upgrades.UpgradeKeepers,
 	mm *module.Manager,
 	configurator module.Configurator,
 ) upgradetypes.UpgradeHandler {
 	return func(ctx sdk.Context, plan upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
-		if err := HandleUpgrade(ctx, rpKeeper, evmKeeper); err != nil {
+		if err := HandleUpgrade(ctx, kk.RpKeeper, kk.EvmKeeper); err != nil {
 			return nil, err
 		}
 		return mm.RunMigrations(ctx, configurator, fromVM)
