+++ app/upgrades/threed/constants.go
@@ -1,4 +1,4 @@
-package foo
+package threed
 
 import (
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
