+++ app/upgrades/threed/doc.go
@@ -1,0 +1,2 @@
+// Package threed (3D) provides a 3D upgrade for the apps Nim+Mande from mainnet (rollapp-evm v2.2.1) to a 3D compatible version.
+package threed
