+++ app/upgrades/threed/upgrade.go
@@ -1,4 +1,4 @@
-package foo
+package threed
 
 import (
 	"fmt"
